+++ WooCommerce/v2/Coupon.cs
@@ -185,26 +185,8 @@ namespace WooCommerceNET.WooCommerce.v2
     }
 
     [DataContract]
-    public class CouponMeta
+    public class CouponMeta : WCObject.MetaData
     {
-        /// <summary>
-        /// Meta ID. 
-        /// read-only
-        /// </summary>
-        [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
-
-        /// <summary>
-        /// Meta key.
-        /// </summary>
-        [DataMember(EmitDefaultValue = false)]
-        public string key { get; set; }
-
-        /// <summary>
-        /// Meta value.
-        /// </summary>
-        [DataMember(EmitDefaultValue = false)]
-        public string value { get; set; }
 
     }
 }