+++ WooCommerce/v2/Customer.cs
@@ -265,26 +265,8 @@ namespace WooCommerceNET.WooCommerce.v2
     }
     
     [DataContract]
-    public class CustomerMeta
+    public class CustomerMeta : WCObject.MetaData
     {
-        /// <summary>
-        /// Meta ID. 
-        /// read-only
-        /// </summary>
-        [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
-
-        /// <summary>
-        /// Meta key.
-        /// </summary>
-        [DataMember(EmitDefaultValue = false)]
-        public string key { get; set; }
-
-        /// <summary>
-        /// Meta value.
-        /// </summary>
-        [DataMember(EmitDefaultValue = false)]
-        public string value { get; set; }
 
     }