+++ WooCommerce/v2/OrderRefund.cs
@@ -70,26 +70,8 @@ namespace WooCommerceNET.WooCommerce.v2
     }
 
     [DataContract]
-    public class OrderRefundMeta
+    public class OrderRefundMeta : WCObject.MetaData
     {
-        /// <summary>
-        /// Meta ID. 
-        /// read-only
-        /// </summary>
-        [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
-
-        /// <summary>
-        /// Meta key.
-        /// </summary>
-        [DataMember(EmitDefaultValue = false)]
-        public string key { get; set; }
-
-        /// <summary>
-        /// Meta value.
-        /// </summary>
-        [DataMember(EmitDefaultValue = false)]
-        public string value { get; set; }
 
     }