+++ WooCommerce/v2/Product.cs
@@ -652,26 +652,9 @@ namespace WooCommerceNET.WooCommerce.v2
     }
 
     [DataContract]
-    public class ProductMeta
+    public class ProductMeta : WCObject.MetaData
     {
-        /// <summary>
-        /// Meta ID. 
-        /// read-only
-        /// </summary>
-        [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
 
-        /// <summary>
-        /// Meta key.
-        /// </summary>
-        [DataMember(EmitDefaultValue = false)]
-        public string key { get; set; }
-
-        /// <summary>
-        /// Meta value.
-        /// </summary>
-        [DataMember(EmitDefaultValue = false)]
-        public object value { get; set; }
     }
 
     [DataContract]